requirements.txt
aiohttp
python-dotenv
redis
pytest
pytest-asyncio
